+++ R/calc_edges.R
@@ -249,14 +249,17 @@ calc_edges <- function() {
 #'    netID <- netID + 1
 #'  }
 #'}
+#'
+#'calcCatchmArea_assignNetID(dt <- dt.streams, id=i, netID)
 
 calcCatchmArea_assignNetID <- function(dt, id, net_ID){
+  print(id)
   if(dt[stream == id, prev_str01,] == 0){  # check only one of prev01 and prev02 because they are always both 0
     dt[stream == id, total_area := area]
     dt[stream == id, netID := net_ID]
   } else {
     a1 <- calcCatchmArea_assignNetID(dt, dt[stream == id, prev_str01], net_ID)
-    a2 <- calcCatchmArea_assignNetID(dt, dt[stream == id, prev_str02] ,net_ID)
+    a2 <- calcCatchmArea_assignNetID(dt, dt[stream == id, prev_str02], net_ID)
     dt[stream == id, total_area := a1 + a2 + dt[stream == id, area]]
     dt[stream == id, netID := net_ID]
   }